internal: refactor remove this semicolon diagnostics
