--- crates/ide/src/diagnostics/fixes/create_field.rs
@@ -1,1 +1,0 @@
-
